+++ app/build.gradle
@@ -63,7 +63,7 @@ dependencies {
     implementation("io.coil-kt:coil:1.2.1")
     implementation "androidx.legacy:legacy-support-v4:1.0.0"
 
-    //testImplementation 'junit:junit:4.+'
-    //androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    //androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    testImplementation 'junit:junit:4.+'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 }
