+++ app/src/main/java/com/x5s/connecting_your_app_to_the_internet/api/Repository.kt
@@ -4,5 +4,4 @@ class Repository(private val apiService: ApiService) {
 
     suspend fun getResults(results: String)  = apiService.getResults(results)
 
-    suspend fun getResultDetails(id:String) = apiService.getResultDetails(id)
 }
