+++ app/src/main/java/com/x5s/connecting_your_app_to_the_internet/api/Result.kt
@@ -3,9 +3,6 @@ package com.x5s.connecting_your_app_to_the_internet.api
 import com.squareup.moshi.Json
 
 data class Result(
-
-        @Json(name = "id")
-        val resultId:String,
         @Json(name = "name")
         val resultName:String,
         @Json(name = "status")
@@ -14,9 +11,6 @@ data class Result(
         val resultSpecies:String,
         @Json(name = "image")
         val resultImage:String
-
 )
 
-//"strDrink": "1-900-FUK_MEUP",
-//"strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/uxywyw1468877224.jpg",
-//"idDrink": "15395"
+
