--- app/src/main/java/com/x5s/connecting_your_app_to_the_internet/api/ResultDetailResponse.kt
@@ -1,5 +1,0 @@
-package com.x5s.connecting_your_app_to_the_internet.api
-
-data class ResultDetailResponse(
-    val results:List<ResultDetail>
-)
