--- app/src/main/java/com/x5s/connecting_your_app_to_the_internet/fragments/detail/DetailFragment.kt
@@ -1,54 +1,0 @@
-package com.x5s.connecting_your_app_to_the_internet.fragments.detail
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import coil.load
-import com.x5s.connecting_your_app_to_the_internet.api.Api
-import com.x5s.connecting_your_app_to_the_internet.api.Repository
-import com.x5s.connecting_your_app_to_the_internet.api.ResultDetail
-import com.x5s.connecting_your_app_to_the_internet.databinding.FragmentDetailBinding
-import com.x5s.connecting_your_app_to_the_internet.fragments.main.MainFragment.Companion.ID_ARGS
-
-class DetailFragment: Fragment() {
-
-    private var binding : FragmentDetailBinding? = null
-    private val resultDetail = mutableListOf<ResultDetail>()
-
-    private val viewModel: DetailFragmentViewModel by lazy{
-        ViewModelProvider(
-                this,
-                DetailViewModelFactory(Repository(Api.apiService))
-        ).get(DetailFragmentViewModel::class.java)
-    }
-
-    override fun onCreateView(
-            inflate: LayoutInflater, container: ViewGroup?,
-            savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = FragmentDetailBinding.inflate(layoutInflater)
-        return binding?.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?){
-        super.onViewCreated(view, savedInstanceState)
-        //viewModel.getResultDetails()
-        var id = ""
-        arguments?.let{
-            id = it.getString(ID_ARGS).toString()
-        }
-        viewModel.getResult(id)
-        viewModel.resultDetailLiveData.observe(viewLifecycleOwner, Observer{
-           resultDetail.addAll(it)
-        })
-
-        binding?.nameTv?.text = resultDetail[0].resultName
-        binding?.imageIv?.load(resultDetail[0].resultImage)
-        activity?.title = resultDetail[0].resultName
-    }
-}
