--- app/src/main/res/navigation/main_nav.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/main_nav"
-    xmlns:tools="http://schemas.android.com/tools"
-    app:startDestination="@id/mainFragment">
-
-    <fragment
-        android:id="@+id/mainFragment"
-        android:name="com.x5s.connecting_your_app_to_the_internet.fragments.main.MainFragment"
-        android:label="fragment_main"
-        tools:layout="@layout/fragment_main">
-
-        <action
-            android:id="@+id/action_mainFragment_to_detailFragment"
-            app:destination="@+id/detailFragment"/>
-    </fragment>
-    <fragment
-        android:id="@+id/detailFragment"
-        android:name="com.x5s.connecting_your_app_to_the_internet.fragments.detail.DetailFragment"
-        android:label="fragment_detail"
-        tools:layout="@layout/fragment_detail">
-
-        <argument
-           android:name="id"
-           app:argType="string" />
-    </fragment>
-
-</navigation>
